services/embedding-service/pyproject.toml
[project]
name = "embedding-service"
version = "1.0.0"
description = "Embedding model microservice for Verifiable RAG system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    # ML / Embeddings
    "sentence-transformers>=2.2.2",
    "torch>=2.1.0",
    # Database (optional - for dynamic config)
    "sqlalchemy[asyncio]>=2.0.25",
    "aiomysql>=0.2.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "cryptography>=46.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",  # For testing
    "ruff>=0.1.0",
]

# For when shared is published as a package
# shared = ["verifiable-rag-shared>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# =============================================================================
# UV Configuration for Local Development
# =============================================================================
# This allows importing the shared package from the monorepo
# Run: uv sync --all-extras

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

# Uncomment this to use the local shared package:
# [tool.uv.sources]
# shared = { path = "../../../shared" }

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "W", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
